move auth view
