+++ app/view/auth.go
